backend/app/workers/tasks.py
from celery import Celery
from sqlalchemy.orm import Session
from .. import models, database
from ..config import settings
import json
import requests
from typing import Dict, Any

# Initialize Celery
celery = Celery(
    "agentic_ai",
    broker=settings.REDIS_URL,
    backend=settings.REDIS_URL
)

@celery.task
def execute_action(action_id: int, action_type: str, details: str) -> Dict[str, Any]:
    """Execute an action asynchronously."""
    # Get database session
    db = next(database.get_db())
    
    try:
        # Get action from database
        action = db.query(models.Action).filter(models.Action.id == action_id).first()
        if not action:
            return {"status": "error", "message": "Action not found"}
        
        # Parse action details
        action_details = json.loads(details)
        
        # Execute action based on type
        if action_type == "email":
            result = send_email(action_details)
        elif action_type == "calendar":
            result = create_calendar_event(action_details)
        elif action_type == "ticket":
            result = create_ticket(action_details)
        else:
            result = {"status": "error", "message": "Unknown action type"}
        
        # Update action status
        action.status = "completed" if result["status"] == "success" else "failed"
        action.details = json.dumps(result)
        db.commit()
        
        return result
    
    except Exception as e:
        # Update action status to failed
        action.status = "failed"
        action.details = json.dumps({"status": "error", "message": str(e)})
        db.commit()
        return {"status": "error", "message": str(e)}
    
    finally:
        db.close()

def send_email(details: Dict[str, Any]) -> Dict[str, Any]:
    """Send an email using the configured email service."""
    # TODO: Implement email sending logic
    return {"status": "success", "message": "Email sent successfully"}

def create_calendar_event(details: Dict[str, Any]) -> Dict[str, Any]:
    """Create a calendar event using the configured calendar service."""
    # TODO: Implement calendar event creation logic
    return {"status": "success", "message": "Calendar event created successfully"}

def create_ticket(details: Dict[str, Any]) -> Dict[str, Any]:
    """Create a ticket using the configured ticketing service."""
    # TODO: Implement ticket creation logic
    return {"status": "success", "message": "Ticket created successfully"} 