backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
langchain==0.0.335
openai==1.3.5
weaviate-client==3.25.2
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1 